requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
websockets>=13.0
google-genai==0.3.0
python-dotenv==1.0.1
google-adk